backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pypdf==4.0.1
tenacity==8.2.3
google-generativeai==0.4.0
python-dotenv
